Split monomorphize inference
